requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
praw==7.7.1
Flask==3.0.0
python-dateutil==2.8.2
lxml==5.1.0
pytz==2024.1
shortuuid==1.0.13
Flask-Session==0.8.0
requests-oauthlib==1.3.1
GitPython==3.1.44
